+++ lib/deribit.rb
@@ -1,10 +1,12 @@
-require 'httparty'
+require 'faraday'
+require 'faraday_middleware'
 require 'hashie'
 
 require 'deribit/version'
 require 'deribit/client'
 
+# Root module
 module Deribit
+  # Base error class
   class Error < StandardError; end
-  # Your code goes here...
 end