+++ test/minitest_helper.rb
@@ -1,0 +1,6 @@
+require 'simplecov'
+SimpleCov.start
+
+require 'bundler/setup'
+require 'deribit-api'
+require 'minitest/autorun'