+++ CHANGELOG.md
@@ -5,6 +5,13 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- HTTP public endpoints
+### Changed
+- use Minitest instead of RSpec
+- use Faraday instead of HTTParty
+
 ## [0.0.1] - 2019-01-04
 ### Added
 - Initial gem release